# backend/gateway_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway банковского приложения. Единая точка входа для клиентских запросов.",
        "title": "Gateway Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/accounts/{account_id}/deposit": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Вносит средства на указанный счёт. Поддерживается только для RUB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Пополнение счёта",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Сумма пополнения",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.AmountPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Транзакция пополнения создана",
                        "schema": {
                            "$ref": "#/definitions/schemas.TransactionDTO"
                        }
                    },
                    "400": {
                        "description": "Неверная сумма или валюта",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{account_id}/transactions": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает список всех операций по конкретному счёту с поддержкой фильтрации по типу.\nРезультаты отсортированы по убыванию даты (сначала новые).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "История транзакций по счёту",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по типу (deposit|withdrawal|transfer|exchange)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Максимальное количество записей",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список транзакций",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.TransactionDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{account_id}/withdraw": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Списывает средства с указанного счёта.\nПроверяет наличие достаточного баланса и лимиты.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Снятие средств",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Сумма снятия",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.AmountPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Транзакция снятия создана",
                        "schema": {
                            "$ref": "#/definitions/schemas.TransactionDTO"
                        }
                    },
                    "400": {
                        "description": "Недостаточно средств",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/metals/rates": {
            "get": {
                "description": "Возвращает текущие банковские цены на золото (XAU), серебро (XAG), платину (XPT) и палладий (XPD).\nЦены указаны за 1 грамм в выбранной валюте.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metals"
                ],
                "summary": "Котировки драгоценных металлов",
                "parameters": [
                    {
                        "type": "string",
                        "default": "RUB",
                        "description": "Валюта цены (ISO 4217)",
                        "name": "base",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список актуальных котировок",
                        "schema": {
                            "$ref": "#/definitions/schemas.MetalRatesResponse"
                        }
                    },
                    "502": {
                        "description": "Внешний источник цен недоступен",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/transactions": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Унифицированный эндпоинт для любого типа транзакций (deposit, withdrawal, transfer).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Универсальная финансовая операция",
                "parameters": [
                    {
                        "description": "Детали операции",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция успешно создана",
                        "schema": {
                            "$ref": "#/definitions/schemas.TransactionDTO"
                        }
                    },
                    "400": {
                        "description": "Невалидные данные запроса",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/transfers": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Перевод между двумя счетами (своими или другим клиентам).\nПри межвалютном переводе используется курс из currency_service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Перевод средств",
                "parameters": [
                    {
                        "description": "Отправитель, получатель, сумма",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Перевод успешно инициирован",
                        "schema": {
                            "$ref": "#/definitions/schemas.TransactionDTO"
                        }
                    },
                    "400": {
                        "description": "Ошибка перевода (недостаточно средств или неверный счет)",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.AmountPayload": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "description": "Сумма операции (положительное число)",
                    "type": "number",
                    "minimum": 0.01,
                    "example": 1000.5
                }
            }
        },
...
        "schemas.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "type"
            ],
            "properties": {
                "account_id": {
                    "description": "ID основного счёта (UUID)",
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "amount": {
                    "description": "Сумма операции",
                    "type": "number",
                    "minimum": 0.01,
                    "example": 1000
                },
                "description": {
                    "description": "Описание/комментарий",
                    "type": "string",
                    "example": "Пополнение через банкомат"
                },
                "to_account_id": {
                    "description": "ID счёта получателя (только для переводов)",
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "type": {
                    "description": "Вид операции",
                    "type": "string",
                    "enum": [
                        "deposit",
                        "withdrawal",
                        "transfer"
                    ],
                    "example": "deposit"
                }
            }
        },
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Подробное сообщение об ошибке для разработчика и пользователя",
                    "type": "string",
                    "example": "Неверный PIN-код или формат данных"
                }
            }
        },
...
        "schemas.MetalPriceDTO": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "Валюта стоимости",
                    "type": "string",
                    "example": "RUB"
                },
                "metal": {
                    "description": "Код металла (XAU - золото, XAG - серебро, XPT - платина, XPD - палладий)",
                    "type": "string",
                    "enum": [
                        "XAU",
                        "XAG",
                        "XPT",
                        "XPD"
                    ],
                    "example": "XAU"
                },
                "price": {
                    "description": "Текущая стоимость за единицу (грамм)",
                    "type": "number",
                    "example": 5850.4
                },
                "unit": {
                    "description": "Единица измерения (всегда 'gram')",
                    "type": "string",
                    "example": "gram"
                }
            }
        },
        "schemas.MetalRatesResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "description": "Валюта оценки",
                    "type": "string",
                    "example": "RUB"
                },
                "rates": {
                    "description": "Список цен на поддерживаемые металлы",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.MetalPriceDTO"
                    }
                },
                "timestamp": {
                    "description": "Время формирования котировок",
                    "type": "string",
                    "example": "2023-10-27T10:00:00Z"
                }
            }
        },
...
        "schemas.TransactionDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "description": "ID счёта, по которому прошла операция",
                    "type": "string",
                    "format": "uuid",
                    "example": "acc-uuid"
                },
                "amount": {
                    "description": "Сумма операции",
                    "type": "number",
                    "example": 500
                },
                "currency": {
                    "description": "Валюта операции",
                    "type": "string",
                    "example": "RUB"
                },
                "description": {
                    "description": "Назначение платежа или техническое описание",
                    "type": "string",
                    "example": "Перевод другу"
                },
                "id": {
                    "description": "Уникальный идентификатор операции (UUID)",
                    "type": "string",
                    "format": "uuid",
                    "example": "tx-uuid"
                },
                "status": {
                    "description": "Текущий статус проводки",
                    "type": "string",
                    "enum": [
                        "pending",
                        "completed",
                        "failed",
                        "blocked"
                    ],
                    "example": "completed"
                },
                "timestamp": {
                    "description": "Точное время совершения операции (ISO 8601)",
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-10-27T10:05:00Z"
                },
                "to_account_id": {
                    "description": "ID счёта контр-агента (только для переводов и обмена)",
                    "type": "string",
                    "format": "uuid",
                    "example": "target-acc-uuid"
                },
                "type": {
                    "description": "Тип финансового действия",
                    "type": "string",
                    "enum": [
                        "deposit",
                        "withdrawal",
                        "transfer",
                        "exchange"
                    ],
                    "example": "transfer"
                }
            }
        },
        "schemas.TransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "to_account_id"
            ],
            "properties": {
                "amount": {
                    "description": "Сумма перевода",
                    "type": "number",
                    "minimum": 0.01,
                    "example": 500
                },
                "description": {
                    "description": "Назначение платежа (комментарий для получателя)",
                    "type": "string",
                    "example": "Перевод другу"
                },
                "to_account_id": {
                    "description": "ID счёта получателя (UUID)",
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OnboardingToken": {
            "description": "Токен регистрации, полученный на Шаге 0 (Start Onboarding)",
            "type": "apiKey",
            "name": "X-Onboarding-Token",
            "in": "header"
        },
        "SessionToken": {
            "description": "Сессионный токен, полученный при авторизации",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    },
...
}